pyproject.toml
[project]
name = "autopilot"
version = "0.1.0"
description = "Autonomous project session orchestrator for Claude Code"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "pyyaml>=6.0",
    "claude-agent-sdk>=0.1.0",
]

[project.scripts]
autopilot = "autopilot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autopilot"]

[tool.hatch.build.targets.wheel.force-include]
"src/autopilot/agents" = "autopilot/agents"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
